+++ CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to this project will be documented here. The format is based
 - Added "Show All" to the "Display" menu of a component (previously: would be available if the right-clicks
   nothing, #527)
 - Added a "Show/Hide all of TYPE" option to the "Display" menu of a component in the editor (#528)
+- The status of the underlying simulator is now presented in the simulator UI's toolbar (#538)
 
 ## [0.3.2] - 2023/01/09